Example of broken out service for maintainability and testability
