+++ static/js/greenwallet/services.js
@@ -1,5 +1,9 @@
-angular.module('greenWalletServices', [])
-.factory('focus', ['$rootScope', '$timeout', 'cordovaReady', function ($rootScope, $timeout, cordovaReady) {
+var greenWalletServices = angular.module('greenWalletServices', []);
+module.exports = greenWalletServices;
+
+require('./services/index')(greenWalletServices);
+
+greenWalletServices.factory('focus', ['$rootScope', '$timeout', 'cordovaReady', function ($rootScope, $timeout, cordovaReady) {
    return function(name) {
         $timeout(function (){
             $rootScope.$broadcast('focusOn', name);
@@ -1914,100 +1918,6 @@ angular.module('greenWalletServices', [])
         }).result;
     };
     return walletsService;
-}]).factory('notices', ['$rootScope', '$timeout', function($rootScope, $timeout) {
-    var notices = $rootScope.notices = [];
-    var noticesService = {};
-    noticesService.makeError = makeError;
-    noticesService.makeNotice = makeNotice;
-    noticesService.setLoadingText = setLoadingText;
-
-    var errorMessages = {
-        notenoughmoney: gettext("Not enough money, you need ${missing_satoshis} more ${unit} to cover the transaction and fee")
-    };
-
-    return noticesService;
-
-    function makeError ($scope, error) {
-        // this is the error object directly from an API call
-        if (error.error !== 'com.greenaddress.error') {
-            console.warn('Non-greenaddress error passed to makeError');
-            return noticesService.makeNotice('error', error.message || error.msg || error);
-        }
-        var code = getErrorCode(error);
-        var message = error.args[1];
-        var args = parseArgs(error.args[2] || {});
-
-        if (code in errorMessages) {
-            message = errorMessages[code];
-        }
-        message = decorateError(message, args);
-
-        return noticesService.makeNotice('error', message);
-
-        // we want the scope
-        function parseArgs (args) {
-            var div = {'BTC': 1, 'mBTC': 1000, 'µBTC': 1000000, 'bits':1000000};
-            args.unit = $scope.wallet.unit;
-
-            // special parsing rules
-            if (args.missing_satoshis) {
-                args.missing_satoshis = Math.round(args.missing_satoshis * div[$scope.wallet.unit]) / 100000000;
-            }
-            return args;
-        }
-    }
-    function decorateError (message, args) {
-        Object.keys(args).forEach(function (argName) {
-            message = message.replace('${' + argName + '}', args[argName]);
-        });
-        return message;
-    }
-    function getErrorCode (error) {
-        return error.args[0].substr(error.args[0].indexOf('#')+1);
-    }
-
-    function makeNotice (type, msg, timeout) {
-        if (msg == null || msg.length == 0)
-            return;
-
-        console.log(msg);
-        var is_chrome_app = window.chrome && chrome.storage;
-        if (is_chrome_app) {
-            var opt = {
-                type: "basic",
-                title: "GreenAddress Notification",
-                message: msg,
-                iconUrl: BASE_URL + "/static/img/logos/logo-greenaddress.png"
-            };
-
-            chrome.notifications.create("", opt);
-        }
-
-        var data = {
-            type: type,
-            msg: msg
-        };
-        notices.push(data);
-
-        if (timeout == null)
-            timeout = 5000;
-
-        if (timeout > 0) {
-            $timeout(function() {
-                for (var i = 0; i < notices.length; ++i) {
-                    if (notices[i] === data) {
-                        notices.splice(i, 1);
-                    }
-                }
-            }, timeout);
-        }
-    }
-    function setLoadingText (text, ifNotSet) {
-        if (!ifNotSet || !$rootScope.loading_text) {
-            $rootScope.loading_text = text;
-        }
-    }
-
 }]).factory('tx_sender', ['$q', '$rootScope', 'cordovaReady', '$http', 'notices', 'gaEvent', '$location', 'autotimeout', 'device_id', 'btchip',
         function($q, $rootScope, cordovaReady, $http, notices, gaEvent, $location, autotimeout, device_id, btchip) {
     var txSenderService = {};
